requirements.txt
geopandas==0.14.*
hydra-core==1.3.*
jupyterlab==4.1.*
matplotlib==3.8.*
numpy==1.26.*
pandas==2.2.*
pyarrow==15.0.*
rasterio==1.3.*
scikit-learn==1.4.*
scipy==1.12.*
segmentation-models-pytorch==0.3.3
torch==2.2.*
